+++ src/raytracer/accelerator/aabb.h
@@ -19,7 +19,7 @@ namespace raytracer
 
 		[[nodiscard]] Point3 maximum() const;
 
-		[[nodiscard]] bool hit(const Ray& r, float t_min, float t_max) const;
+		[[nodiscard]] bool hit(const Ray& r, double t_min, double t_max) const;
 
 		static Aabb surroundingBox(const Aabb& box0, const Aabb& box1);
