+++ src/raytracer/cameras/camera.h
@@ -8,17 +8,17 @@ namespace raytracer
 
 	struct CameraSpecification
 	{
-		float vfov = 90.0;
-		float aspect_ratio = 16.0 / 9.0;
-		float aperture = 0.0;
-		float focusDistance = 10.0;
+		double vfov = 90.0;
+		double aspect_ratio = 16.0 / 9.0;
+		double aperture = 0.0;
+		double focusDistance = 10.0;
 
 		Point3 lookFrom = { 0, 0, 0 };
 		Point3 lookAt = { 0, 0, -1 };
 		Vector3 vup = { 0, 1, 0 };
 
-		float time0 = 0; // shutter opens
-		float time1 = 0; // shutter closes
+		double time0 = 0; // shutter opens
+		double time1 = 0; // shutter closes
 	};
 
 	class Camera
@@ -26,21 +26,21 @@ namespace raytracer
 	public:
 		explicit Camera(const CameraSpecification& spec);
 
-		[[nodiscard]] Ray getRay(float u, float v) const;
+		[[nodiscard]] Ray getRay(double u, double v) const;
 
 		void onResize(uint32_t width, uint32_t height);
 
-		float vfov() const
+		double vfov() const
 		{
 			return m_specification.vfov;
 		}
 
-		float aperture() const
+		double aperture() const
 		{
 			return m_specification.aperture;
 		}
 
-		float focusDistance() const
+		double focusDistance() const
 		{
 			return m_specification.focusDistance;
 		}
@@ -55,12 +55,12 @@ namespace raytracer
 			return m_specification.lookAt;
 		}
 
-		float time0() const
+		double time0() const
 		{
 			return m_specification.time0;
 		}
 
-		float time1() const
+		double time1() const
 		{
 			return m_specification.time1;
 		}
@@ -79,7 +79,7 @@ namespace raytracer
 		Vector3 m_vertical{};
 		Vector3 m_u{}, m_v{}, m_w{};
 
-		float m_lensRadius;
+		double m_lensRadius;
 
 		CameraSpecification m_specification;
 	};
