+++ src/raytracer/hittable/box.h
@@ -20,9 +20,9 @@ namespace raytracer
 
 		Box(const Point3& p0, const Point3& p1, const ref<Material>& ptr);
 
-		bool hit(const Ray& r, float t_min, float t_max, hit_record& rec) const override;
+		bool hit(const Ray& r, double t_min, double t_max, hit_record& rec) const override;
 
-		bool boundingBox(float time0, float time1, Aabb& output_box) const override;
+		bool boundingBox(double time0, double time1, Aabb& output_box) const override;
 
 	public:
 		Point3 box_min{};
